Move model.playlist package out to model
